# guardian/guardian-api-client/swagger.yaml
openapi: 3.0.0
info:
  title: 'Guardian'
  description: 'The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation.'
  version: '2.10.0'
  contact:
    name: 'API developer'
    url: 'https://envisionblockchain.com'
    email: 'info@envisionblockchain.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: /api/v1
    description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: 'http'
      scheme: 'bearer'
      bearerFormat: 'JWT'
  schemas:
...
    ExportPolicy:
      type: object
      required:
        - name
        - version
        - messageId
      properties:
        name:
          type: string
        version:
          type: string
        tokens:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    CommonSettings:
      type: object
      properties:
        operatorId:
          type: string
        operatorKey:
          type: string
        nftApiKey:
          deprecated: true
          type: string
        ipfsStorageApiKey:
          type: string
...
paths:
...
  /policies/{policyId}/export/file:
    get:
      tags:
        - policies
      description: Returns a zip file containing the published policy and all associated artifacts, i.e. schemas and VCs. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: policyId
          schema:
            type: string
          required: true
          description: Selected policy ID.
      security:
        - bearerAuth: []
      summary: Return policy and its artifacts in a zip file format for the specified policy.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPolicy'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /ipfs/file/{cid}:
    get:
      tags:
        - ipfs
      description: Get file from ipfs.
      summary: Get file from ipfs.
      parameters:
        - in: path
          name: cid
          schema:
            type: string
          required: true
          description: File CID.
      security:
        - bearerAuth: []
      responses:
        201:
          description: Successful operation.
          content:
            binary/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /settings:
    get:
      tags:
        - settings
      description: Returns current settings. For users with the Standard Registry role only.
      security:
        - bearerAuth: []
      summary: Returns current settings.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonSettings'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - settings
      description: Set settings. For users with the Standard Registry role only.
      security:
        - bearerAuth: []
      summary: Set settings.
      responses:
        200:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonSettings'
...
  /modules/{uuid}/export/file:
    get:
      tags:
        - modules
      description: Returns a zip file containing the published module and all associated artifacts, i.e. schemas and VCs. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
          required: true
          description: Selected module ID.
      security:
        - bearerAuth: []
      summary: Return module and its artifacts in a zip file format for the specified module.
      responses:
        200:
          description: Successful operation. Response zip file
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...